controller and routes setup

+++ src/utils/cloudinary.js
@@ -1,50 +1,28 @@
-import { v2 as cloudinary } from 'cloudinary';
+import { v2 as cloudinary } from "cloudinary";
 import fs from "fs";
 
-    // Configuration
-    cloudinary.config({ 
-        cloud_name: process.env.CLOUD_NAME, 
-        api_key: process.env.CLOUDINARY_API_KEY, 
-        api_secret:process.env.CLOUDINARY_API_SECRET
+cloudinary.config({
+  cloud_name: process.env.CLOUD_NAME,
+  api_key: process.env.CLOUDINARY_API_KEY,
+  api_secret: process.env.CLOUDINARY_API_SECRET,
+});
+
+const uploadOnCloudinary = async (localFilePath) => {
+  try {
+    if (!localFilePath) return null;
+
+    const response = await cloudinary.uploader.upload(localFilePath, {
+      resource_type: "auto",
     });
-    
-    // Upload an image
-     const uploadOnCloudinary = async(localFilePath)=>{
-        try{
-            if(!localFilePath) return null;
-            //upload the file on cloudinary
-          let response=await cloudinary.uploader.upload(
-           localFilePath, {
-                resource_type:"auto"
-           })
-           //file has be successfully uploaded
-           console.log("File is uploaded on clodinary")
-           console.log(response.url)
-           return response;
-        }
-       catch(error){
-        fs.unlinkSync(localFilePath)
-           console.log(error);
-       }
+
+    fs.unlinkSync(localFilePath); // local file delete
+    return response;
+  } catch (error) {
+    if (localFilePath && fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
     }
-    
-    console.log(uploadOnCloudinary);
-    
-    // Optimize delivery by resizing and applying auto-format and auto-quality
-    const optimizeUrl = cloudinary.url('shoes', {
-        fetch_format: 'auto',
-        quality: 'auto'
-    });
-    
-    console.log(optimizeUrl);
-    
-    // Transform the image: auto-crop to square aspect_ratio
-    const autoCropUrl = cloudinary.url('shoes', {
-        crop: 'auto',
-        gravity: 'auto',
-        width: 500,
-        height: 500,
-    });
-    
-    console.log(autoCropUrl);    
+    throw error;
+  }
+};
+
 export { uploadOnCloudinary };
